# engine/docs/swagger.yaml
definitions:
  AgentProtocol:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - AgentProtocol_UNKNOWN
    - AgentProtocol_GRPC
    - AgentProtocol_HTTP
  AgentResponse:
    properties:
      address:
        type: string
      name:
        type: string
      port:
        type: string
      protocol:
        $ref: '#/definitions/AgentProtocol'
      supportedTasks:
        items:
          type: string
        type: array
      version:
        type: string
    type: object
  DecisionCase:
    properties:
      condition:
        type: string
      description:
        type: string
      name:
        type: string
      nextStepId:
        type: string
    required:
    - condition
    - nextStepId
    type: object
  DecisionConfig:
    properties:
      cases:
        items:
          $ref: '#/definitions/DecisionCase'
        type: array
      joinStepId:
        type: string
    required:
    - cases
    - joinStepId
    type: object
  ForkBranch:
    properties:
      description:
        type: string
      name:
        type: string
      nextStepId:
        type: string
    required:
    - nextStepId
    type: object
  ForkConfig:
    properties:
      branches:
        items:
          $ref: '#/definitions/ForkBranch'
        type: array
      joinStepId:
        type: string
    required:
    - branches
    - joinStepId
    type: object
  JoinConfig:
    properties:
      incomingStepIds:
        items:
          type: string
        type: array
      nextStepId:
        type: string
    required:
    - incomingStepIds
    type: object
  StepDefinitionParameter:
    properties:
      type:
        $ref: '#/definitions/StepParameterType'
      value: {}
    required:
    - type
    - value
    type: object
  StepDefinitionParameters:
    additionalProperties:
      $ref: '#/definitions/StepDefinitionParameter'
    type: object
  StepInstance:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      errorMessage:
        type: string
      id:
        type: string
      input:
        $ref: '#/definitions/UnknownJson'
      output:
        $ref: '#/definitions/UnknownJson'
      progress:
        type: number
      startedAt:
        type: string
      status:
        $ref: '#/definitions/StepStatus'
      stepId:
        type: string
      updatedAt:
        type: string
      workflowDefinitionId:
        type: string
      workflowInstanceId:
        type: string
    type: object
  StepParameterType:
    enum:
    - constant
    - workflow
    - taskOutput
    type: string
    x-enum-varnames:
    - StepParameterTypeConstant
    - StepParameterTypeWorkflow
    - StepParameterTypeTaskOutput
  StepStatus:
    enum:
    - pending
    - running
    - completed
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - StepStatusPending
    - StepStatusRunning
    - StepStatusCompleted
    - StepStatusFailed
    - StepStatusCancelled
  StepType:
    enum:
    - task
    - workflow
    - wait
    - decision
    - fork
    - join
    type: string
    x-enum-varnames:
    - StepTypeTask
    - StepTypeWorkflow
    - StepTypeWait
    - StepTypeDecision
    - StepTypeFork
    - StepTypeJoin
  TaskConfig:
    properties:
      nextStepId:
        type: string
      taskDefinitionId:
        type: string
    required:
    - taskDefinitionId
    type: object
  TaskDefinitionResponse:
    properties:
      description:
        type: string
      id:
        type: string
      inputParameters:
        $ref: '#/definitions/UnknownJson'
      name:
        type: string
      outputParameters:
        $ref: '#/definitions/UnknownJson'
    type: object
  UnknownJson:
    additionalProperties: true
    type: object
  UpdateWorkflowDefRequest:
    properties:
      description:
        type: string
      name:
        type: string
      steps:
        items:
          $ref: '#/definitions/WorkflowStepDefinition'
        type: array
    type: object
  WaitConfig:
    properties:
      durationSeconds:
        $ref: '#/definitions/StepDefinitionParameter'
      nextStepId:
        type: string
    required:
    - durationSeconds
    type: object
  WorkflowConfig:
    properties:
      nextStepId:
        type: string
      workflowDefinitionId:
        type: string
    required:
    - workflowDefinitionId
    type: object
  WorkflowDefinition:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      inputParameters:
        items:
          $ref: '#/definitions/WorkflowParameterDefinition'
        type: array
      isDraft:
        type: boolean
      isEnabled:
        type: boolean
      metadata:
        $ref: '#/definitions/UnknownJson'
      name:
        type: string
      outputParameters: {}
      steps:
        items:
          $ref: '#/definitions/WorkflowStepDefinition'
        type: array
      updatedAt:
        type: string
      version:
        type: string
    required:
    - createdAt
    - id
    - isDraft
    - isEnabled
    - name
    - steps
    - updatedAt
    - version
    type: object
  WorkflowInstance:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      errorMessage:
        type: string
      id:
        type: string
      input:
        $ref: '#/definitions/UnknownJson'
      output:
        $ref: '#/definitions/UnknownJson'
      startedAt:
        type: string
      status:
        $ref: '#/definitions/WorkflowStatus'
      updatedAt:
        type: string
      workflowDefinitionId:
        type: string
    type: object
  WorkflowParameterDefinition:
    properties:
      default: {}
      description:
        type: string
      metadata:
        $ref: '#/definitions/UnknownJson'
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
  WorkflowStatus:
    enum:
    - PENDING
    - RUNNING
    - COMPLETED
    - FAILED
    - CANCELLED
    type: string
    x-enum-varnames:
    - WorkflowStatusPending
    - WorkflowStatusRunning
    - WorkflowStatusCompleted
    - WorkflowStatusFailed
    - WorkflowStatusCancelled
  WorkflowStepDefinition:
    properties:
      decisionConfig:
        $ref: '#/definitions/DecisionConfig'
      description:
        type: string
      forkConfig:
        $ref: '#/definitions/ForkConfig'
      joinConfig:
        $ref: '#/definitions/JoinConfig'
      metadata:
        $ref: '#/definitions/UnknownJson'
      name:
        type: string
      parameters:
        $ref: '#/definitions/StepDefinitionParameters'
      stepDefinitionId:
        type: string
      taskConfig:
        $ref: '#/definitions/TaskConfig'
      type:
        $ref: '#/definitions/StepType'
      waitConfig:
        $ref: '#/definitions/WaitConfig'
      workflowConfig:
        $ref: '#/definitions/WorkflowConfig'
    required:
    - name
    - stepDefinitionId
    - type
    type: object
  gin.H:
    additionalProperties: {}
    type: object
info:
  contact: {}
  title: Workflow Engine API
  version: "1.0"
paths:
  /api/agents:
    get:
      description: Get a list of all registered agents
      operationId: getAllAgents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/AgentResponse'
            type: array
      summary: Get all agents
      tags:
      - agents
  /api/agents/{name}:
    get:
      description: Get details of a registered agent by its name
      operationId: getAgentByName
      parameters:
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AgentResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get agent by name
      tags:
      - agents
  /api/step-instances:
    get:
      description: Get all step instances with pagination
      operationId: getAllStepInstances
      parameters:
      - description: Page limit
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/StepInstance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all step instances
      tags:
      - StepInstances
  /api/step-instances/{id}:
    get:
      description: Get a step instance by its ID
      operationId: getStepInstanceById
      parameters:
      - description: Step Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/StepInstance'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get step instance by ID
      tags:
      - StepInstances
  /api/task-definitions:
    get:
      description: Get all task definitions
      operationId: getAllTaskDefinitions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TaskDefinitionResponse'
            type: array
      summary: Get all task definitions
      tags:
      - Task Definitions
  /api/task-definitions/{id}:
    get:
      description: Get task definition by ID
      operationId: getTaskDefinitionById
      parameters:
      - description: Task Definition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TaskDefinitionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get task definition by ID
      tags:
      - Task Definitions
  /api/workflow-definitions:
    get:
      description: Get all workflow definitions
      operationId: getAllWorkflowDefinitions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/WorkflowDefinition'
            type: array
      summary: Get all workflow definitions
      tags:
      - Workflow Definitions
    post:
      consumes:
      - application/json
      description: Create a new workflow definition
      operationId: createWorkflowDefinition
      parameters:
      - description: Workflow Definition
        in: body
        name: definition
        required: true
        schema:
          $ref: '#/definitions/WorkflowDefinition'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/WorkflowDefinition'
      summary: Create a new workflow definition
      tags:
      - Workflow Definitions
  /api/workflow-definitions/{id}:
    delete:
      description: Delete a workflow definition
      operationId: deleteWorkflowDefinition
      parameters:
      - description: Workflow Definition ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete a workflow definition
      tags:
      - Workflow Definitions
    get:
      description: Get a workflow definition by ID
      operationId: getWorkflowDefinitionByID
      parameters:
      - description: Workflow Definition ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/WorkflowDefinition'
      summary: Get a workflow definition by ID
      tags:
      - Workflow Definitions
    put:
      consumes:
      - application/json
      description: Update an existing workflow definition
      operationId: updateWorkflowDefinition
      parameters:
      - description: Workflow Definition ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Workflow Definition
        in: body
        name: definition
        required: true
        schema:
          $ref: '#/definitions/UpdateWorkflowDefRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/WorkflowDefinition'
      summary: Update an existing workflow definition
      tags:
      - Workflow Definitions
  /api/workflow-definitions/{id}/disable:
    patch:
      description: Delete a workflow definition
      operationId: disableWorkflowDefinition
      parameters:
      - description: Workflow Definition ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete a workflow definition
      tags:
      - Workflow Definitions
  /api/workflow-definitions/{id}/enable:
    patch:
      description: Delete a workflow definition
      operationId: enableWorkflowDefinition
      parameters:
      - description: Workflow Definition ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete a workflow definition
      tags:
      - Workflow Definitions
  /api/workflow-definitions/{id}/publish:
    patch:
      description: Publish a workflow definition
      operationId: publishWorkflowDefinition
      parameters:
      - description: Workflow Definition ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Publish a workflow definition
      tags:
      - Workflow Definitions
  /api/workflow-instances:
    get:
      description: Get all workflow instances
      operationId: getAllWorkflowInstances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/WorkflowInstance'
            type: array
      summary: Get all workflow instances
      tags:
      - Workflow Instances
  /api/workflow-instances/{id}:
    get:
      description: Get a workflow instance by ID
      operationId: getWorkflowInstanceByID
      parameters:
      - description: Workflow Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/WorkflowInstance'
      summary: Get a workflow instance by ID
      tags:
      - Workflow Instances
swagger: "2.0"